Refactor ReportDetails
